refactor: Se limitan los tipos que puede tener los datos a validar
(Se mueve el paquete validation de infrastructure -> domain)
