+++ user/infrastructure/handler/handlers.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"log"
 
+	"api-your-accounts/shared/domain/validation"
 	"api-your-accounts/shared/infrastructure/db"
-	"api-your-accounts/shared/infrastructure/validation"
 	"api-your-accounts/user/application"
 	"api-your-accounts/user/domain"
 	"api-your-accounts/user/infrastructure/model"
